更新Pgsql Store

+++ backend/Yaginx.Core/DomainModels/IUserRepository.cs
@@ -2,7 +2,7 @@
 
 public interface IUserRepository
 {
-    Task<int> CountAsync();
+    Task<long> CountAsync();
     Task<User> GetByEmailAsync(string email);
     Task AddAsync(User user);
     Task UpdateAsync(User user);
